Make transformation Pk --> xir using available 'hankel.py' library
Change the way in which Pk --> xir transform is implemented. To do this
I use the 'hankel.py' library by Steven Murray
(https://pypi.python.org/pypi/hankel), more or less similar to what was
done in CosmoPy, but readily available.
I also find the parameters to make the transform fast and accurate.