Change in the way we pass parameters related to pre-computed grid Fourier profile
We now always pass the individual NP arrays to the different functions and classes
instead of the reference to the NpzFile object, as in the latter case, the
objects are not 'pickeable', so that the multi-thread Emcee would fail.
With this new version we can run the fit with threads > 1 without any problem.